+++ riverpie_flutter/lib/addons.dart
@@ -1,3 +1,2 @@
-export 'package:riverpie_flutter/src/addons/action.dart';
 export 'package:riverpie_flutter/src/addons/navigation.dart';
 export 'package:riverpie_flutter/src/addons/snack_bar.dart';